requirements.txt
streamlit>=1.33
pandas>=2.0
boto3>=1.34
plotly>=5.18
python-dotenv>=1.0
